requirements.txt
pandas
sentence-transformers
pinecone-client
python-dotenv
httpx
numpy
openai  
tqdm
torch
scipy  
transformers
groq
